# internal/adapters/primary/http/clan_equipment_costs.go
package http

import (
    "context"
    "net/http"
    "time"

    "github.com/gin-gonic/gin"

    "github.com/dauletkhan/coc/internal/application/usecases"
)

type ClanEquipmentCostsHandler struct {
    uc *usecases.ClanEquipmentCostsUseCase
}

func NewClanEquipmentCostsHandler(uc *usecases.ClanEquipmentCostsUseCase) *ClanEquipmentCostsHandler {
    return &ClanEquipmentCostsHandler{uc: uc}
}

func (h *ClanEquipmentCostsHandler) Register(r *gin.Engine) {
    r.GET("/v1/clans/:tag/hero-equipments/costs", h.get)
}

func (h *ClanEquipmentCostsHandler) get(c *gin.Context) {
    tag := c.Param("tag")
    if tag == "" {
        c.JSON(http.StatusBadRequest, gin.H{"error": "missing tag"})
        return
    }
    nTag := normalizePlayerTag(tag)

    ctx, cancel := context.WithTimeout(c.Request.Context(), 10*time.Second)
    defer cancel()

    res, status, err := h.uc.Execute(ctx, nTag)
    if err != nil {
        if status == 0 {
            status = http.StatusBadGateway
        }
        c.JSON(status, gin.H{"error": err.Error()})
        return
    }
    if status >= 400 {
        c.Status(status)
        return
    }
    c.JSON(http.StatusOK, res)
}