# internal/http/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Custom Clash of Clans API
  version: 1.0.0
  description: |
    A developer-friendly wrapper around the official Clash of Clans API.

    This service focuses on player hero equipment data, including availability and
    cumulative resource (ore) costs per rarity (COMMON, EPIC). Data for ore costs
    is provided via a local catalog and not from the official API.
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: players
    description: Player-centric endpoints
  - name: clans
    description: Clan-centric endpoints
paths:
  /v1/players/{tag}/hero-equipments:
    get:
      tags: [players]
      summary: List player's hero equipments with availability
      description: |
        Returns player's hero equipments detected from the official API response.
        Marks equipments not present on the player as unavailable based on the local
        equipment catalog.
      parameters:
        - name: tag
          in: path
          required: true
          description: Player tag (URL-encoded, e.g. %23ABC123). The API also accepts raw `#ABC123` or `ABC123`.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  playerTag:
                    type: string
                  available:
                    type: array
                    items:
                      $ref: '#/components/schemas/Equipment'
                  unavailable:
                    type: array
                    items:
                      $ref: '#/components/schemas/Equipment'
              examples:
                sample:
                  summary: Example response
                  value:
                    playerTag: "%2382VCYUURR"
                    available:
                      - name: "Rage Vial"
                        level: 6
                        maxLevel: 18
                        available: true
                    unavailable:
                      - name: "Giant Gauntlet"
                        level: 0
                        maxLevel: 0
                        available: false
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
  /v1/players/{tag}/hero-equipments/costs:
    get:
      tags: [players]
      summary: Get cumulative ore spent per equipment and totals
      description: |
        Computes cumulative ore spent (shiny, glowy, starry) up to the player's current
        level for each equipment. Rarity-specific per-level costs are read from the local
        catalog (`data/hero_equipment.json`).
      parameters:
        - name: tag
          in: path
          required: true
          description: Player tag (URL-encoded, e.g. %23ABC123). The API also accepts raw `#ABC123` or `ABC123`.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  playerTag:
                    type: string
                  total:
                    $ref: '#/components/schemas/OreTotals'
                  equipments:
                    type: array
                    items:
                      $ref: '#/components/schemas/EquipmentSpend'
              examples:
                sample:
                  summary: Example response
                  value:
                    playerTag: "%2382VCYUURR"
                    total:
                      shiny: 67640
                      glowy: 4200
                      starry: 110
                    equipments:
                      - name: "Rage Vial"
                        rarity: "COMMON"
                        level: 6
                        spent: { shiny: 3320, glowy: 120, starry: 0 }
                      - name: "Giant Arrow"
                        rarity: "COMMON"
                        level: 18
                        spent: { shiny: 27260, glowy: 1920, starry: 0 }
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
  /v1/clans/{tag}/hero-equipments/costs:
    get:
      tags: [clans]
      summary: Get clan-wide cumulative ore spent per member and totals
      description: |
        Aggregates shiny, glowy, and starry ore spent for each clan member by fetching
        each player's equipments in parallel and summing costs using the local catalog.
        Members are returned sorted by shiny desc, then glowy desc, then starry desc.
      parameters:
        - name: tag
          in: path
          required: true
          description: Clan tag (URL-encoded, e.g. %23ABC123). The API also accepts raw `#ABC123` or `ABC123`.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  clanTag:
                    type: string
                  total:
                    $ref: '#/components/schemas/OreTotals'
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClanMemberSpend'
              examples:
                sample:
                  summary: Example response
                  value:
                    clanTag: "%23CLAN123"
                    total: { shiny: 123456, glowy: 7890, starry: 210 }
                    members:
                      - tag: "%23PLAYER1"
                        name: "Alice"
                        spent: { shiny: 40000, glowy: 1200, starry: 60 }
                      - tag: "%23PLAYER2"
                        name: "Bob"
                        spent: { shiny: 30000, glowy: 1000, starry: 40 }
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
components:
  schemas:
    Equipment:
      type: object
      properties:
        name:
          type: string
        level:
          type: integer
          description: Current player equipment level (0 when unavailable)
        maxLevel:
          type: integer
          description: Reported max level from the official API when available, else 0
        available:
          type: boolean
    OreTotals:
      type: object
      properties:
        shiny:
          type: integer
        glowy:
          type: integer
        starry:
          type: integer
    EquipmentSpend:
      type: object
      properties:
        name:
          type: string
        rarity:
          type: string
          enum: [COMMON, EPIC, UNKNOWN]
        level:
          type: integer
          description: Current player equipment level
        spent:
          $ref: '#/components/schemas/OreTotals'
    ClanMemberSpend:
      type: object
      properties:
        tag:
          type: string
        name:
          type: string
        spent:
          $ref: '#/components/schemas/OreTotals'